#68: Always use int instead of long because of Python 3

+++ edmunds/foundation/testing/testcommand.py
@@ -47,7 +47,7 @@ class TestCommand(Command):
             type_mapping = {
                 'string': str,
                 'int': int,
-                'long': long,
+                'long': int,
                 'float': float,
                 # 'complex': str,
                 # 'choice': str,
